refactor: Dealer 에 있던 수익계산 코드를 Player 및 Players 로 옮김
Dealer 에서 수익계산을 하려면 Players 에서 Player List 를 꺼내게 된다.
만약 Player 객체가 자신의 수익을 계산할 수 있다면
Dealer 는 Players 에 요청을 할 수 있겠지만,
그러면 Dealer 는 Player 에 요청하고 Player 는 Dealer 에 요청하는
이상한 관계가 만들어진다.
그래서 아예 Player 와 Players 가 딜러의 결과계산까지 수행하도록 하였다.
